+++ examples/2qubit_gates.jl
@@ -2,12 +2,12 @@ function decompose_hadamard()
 
     println(">>>>> Hadamard Gate <<<<<")
  
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 3,    
     "elementary_gates" => ["U3_1", "CNot_1_2", "Identity"], 
-    "target_gate" => QCO.get_full_sized_gate("H_1", 2),
+    "target_gate" => QCOpt.get_full_sized_gate("H_1", 2),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
        
@@ -15,21 +15,19 @@ function decompose_hadamard()
     "U3_ϕ_discretization" => [0, π/2],
     "U3_λ_discretization" => [0, π],
     )
-
-    return params
-    
+ 
 end
 
 function decompose_controlled_Z()
 
     println(">>>>> Controlled-Z Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 4,
     "elementary_gates" => ["U3_1", "U3_2", "CNot_1_2", "Identity"], 
-    "target_gate" => QCO.CZGate(),
+    "target_gate" => QCOpt.CZGate(),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
 
@@ -38,38 +36,34 @@ function decompose_controlled_Z()
     "U3_λ_discretization" => -π:π/2:π,
     )
 
-    return params
-    
 end
 
 function decompose_controlled_V()
 
     println(">>>>> Controlled-V Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 7,
     "elementary_gates" => ["H_1", "H_2", "T_1", "T_2", "Tdagger_1", "Tdagger_2", "CNot_1_2", "CNot_2_1", "Identity"],
-    "target_gate" => QCO.CVGate(),
+    "target_gate" => QCOpt.CVGate(),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
     )
 
-    return params
-    
 end
 
 function decompose_controlled_H()
 
     println(">>>>> Controlled-H Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 5,
     "elementary_gates" => ["U3_1", "U3_2", "CNot_1_2", "CNot_2_1", "Identity"], 
-    "target_gate" => QCO.CHGate(),
+    "target_gate" => QCOpt.CHGate(),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
 
@@ -78,20 +72,18 @@ function decompose_controlled_H()
     "U3_λ_discretization" => [0],
     )
 
-    return params
-    
 end
 
 function decompose_controlled_H_with_R()
 
     println(">>>>> Controlled-H with R Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 5,
     "elementary_gates" => ["RY_1", "RY_2", "CNot_1_2", "Identity"], 
-    "target_gate" => QCO.CHGate(),
+    "target_gate" => QCOpt.CHGate(),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
        
@@ -100,20 +92,18 @@ function decompose_controlled_H_with_R()
     "set_cnot_upper_bound" => 2,           
     )
 
-    return params
-    
 end
 
 function decompose_magic()
     
     println(">>>>> M Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 4,
         "elementary_gates" => ["U3_1", "U3_2", "CNot_2_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.MGate(),   
+        "target_gate" => QCOpt.MGate(),   
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
 
@@ -121,21 +111,19 @@ function decompose_magic()
         "U3_ϕ_discretization" => -π:π/2:π,
         "U3_λ_discretization" => -π:π/2:π,                         
         )
-    
-        return params
-    
+     
 end
 
 function decompose_magic_using_CNOT_1_2()
     
     println(">>>>> M Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 5,
         "elementary_gates" => ["U3_1", "U3_2", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.MGate(),
+        "target_gate" => QCOpt.MGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
            
@@ -144,37 +132,34 @@ function decompose_magic_using_CNOT_1_2()
         "U3_λ_discretization" => -π:π/2:π,            
         )
     
-        return params
-    
 end
 
 function decompose_magic_using_SHCnot()
     
     println(">>>>> M gate using S, H and CNOT Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 5,
         "elementary_gates" => ["S_1", "S_2", "H_1", "H_2", "CNot_1_2", "CNot_2_1", "Identity"], 
-        "target_gate" => QCO.MGate(),
+        "target_gate" => QCOpt.MGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
         )
     
-        return params 
 end
 
 function decompose_S()
 
     println(">>>>> S Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 3,
         "elementary_gates" => ["U3_1", "U3_2", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.get_full_sized_gate("S_1", 2),
+        "target_gate" => QCOpt.get_full_sized_gate("S_1", 2),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
            
@@ -182,39 +167,34 @@ function decompose_S()
         "U3_ϕ_discretization" => [-π/2, 0, π/2],
         "U3_λ_discretization" => [0, π],           
         )
-    
-        return params
-    
 end
 
 function decompose_revcnot()
 
     println(">>>>> Reverse CNOT Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 2, 
     "maximum_depth" => 5,
     "elementary_gates" => ["H_1", "H_2", "CNot_1_2", "Identity"],  
-    "target_gate" => QCO.CNotRevGate(),
+    "target_gate" => QCOpt.CNotRevGate(),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",       
     )
 
-    return params
-    
 end
 
 function decompose_revcnot_with_U()
     
     println(">>>>> Reverse CNOT using U3 and CNOT Gates <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 5,    
         "elementary_gates" => ["U3_1", "U3_2", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.CNotRevGate(),
+        "target_gate" => QCOpt.CNotRevGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
 
@@ -223,25 +203,21 @@ function decompose_revcnot_with_U()
         "U3_λ_discretization" => [0],
         )
     
-        return params
-    
 end
 
 function decompose_swap()
 
     println(">>>>> SWAP Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 5,    
         "elementary_gates" => ["CNot_2_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.SwapGate(),
+        "target_gate" => QCOpt.SwapGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",           
         )
-    
-        return params
     
 end
 
@@ -249,12 +225,12 @@ function decompose_W()
 
     println(">>>>> W Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 5,
         "elementary_gates" => ["U3_2", "CNot_2_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.WGate(),
+        "target_gate" => QCOpt.WGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
 
@@ -263,25 +239,21 @@ function decompose_W()
         "U3_λ_discretization" => [0, π/2],
         )
     
-        return params
-    
 end
 
 function decompose_W_using_HCnot()
 
     println(">>>>> W Gate using H and CNOT gates <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 6,
         "elementary_gates" => ["CH_1_2", "CNot_2_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.WGate(),
+        "target_gate" => QCOpt.WGate(),
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
         )
-    
-        return params
     
 end
 
@@ -289,17 +261,15 @@ function decompose_HCoinGate()
 
     println(">>>>> Hadamard Coin gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 10,
         "elementary_gates" => ["Y_1", "Y_2", "Z_1", "Z_2", "T_2", "Tdagger_1", "Sdagger_1", "SX_1", "SXdagger_2", "CNot_2_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.HCoinGate(),  
+        "target_gate" => QCOpt.HCoinGate(),  
         "objective" => "minimize_depth", 
         "decomposition_type" => "exact",
         )
-    
-        return params
     
 end
 
@@ -307,19 +277,17 @@ function decompose_GroverDiffusion_using_HX()
 
     println(">>>>> Grover's Diffusion Operator <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
         
         #Reference: https://arxiv.org/pdf/1804.03719.pdf (Fig 6)
 
         "num_qubits" => 2, 
         "maximum_depth" => 10,
         "elementary_gates" => ["X_1", "X_1xX_2", "H_1xH_2", "X_2", "H_1", "H_2", "CNot_1_2", "Identity"],
-        "target_gate" => QCO.GroverDiffusionGate(),          
+        "target_gate" => QCOpt.GroverDiffusionGate(),          
         "objective" => "minimize_depth",
         "decomposition_type" => "exact",
         )
-    
-        return params
     
 end
 
@@ -327,17 +295,15 @@ function decompose_GroverDiffusion_using_Clifford()
 
     println(">>>>> Grover's Diffusion Operator <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2,
         "maximum_depth" => 6,
         "elementary_gates" => ["X_1", "X_2", "H_1", "H_2", "S_1", "S_2", "T_1", "T_2", "Y_1", "Y_2", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.GroverDiffusionGate(),
+        "target_gate" => QCOpt.GroverDiffusionGate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "exact",
         )
-    
-        return params
     
 end
 
@@ -346,12 +312,12 @@ function decompose_GroverDiffusion_using_U3()
 
     println(">>>>> Grover's Diffusion Operator using U3 gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 10,
         "elementary_gates" => ["U3_1", "CNot_1_2", "Identity"], 
-        "target_gate" => QCO.GroverDiffusionGate(),
+        "target_gate" => QCOpt.GroverDiffusionGate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "exact",
 
@@ -359,38 +325,35 @@ function decompose_GroverDiffusion_using_U3()
         "U3_ϕ_discretization" => -π:π/2:π,
         "U3_λ_discretization" => [0],
         )
-    
-        return params
-    
+
 end
 
-function decompose_iSwapGate()
+function decompose_iSwap()
 
     println(">>>>> iSwap Gate <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 10,
         "elementary_gates" => ["T_1", "T_2", "Tdagger_1", "Tdagger_2", "H_1", "H_2", "CNot_1_2", "CNot_2_1", "Identity"],
-        "target_gate" => QCO.iSwapGate(),
+        # "elementary_gates" => ["S_1", "S_2", "Sdagger_1", "Sdagger_2", "H_1", "H_2", "CNot_1_2", "CNot_2_1", "Identity"],
+        "target_gate" => QCOpt.iSwapGate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "exact",                           
         )
-    
-        return params
-    
+
 end
 
 function decompose_qft2_using_R()
     println(">>>>> QFT2 Gate using R and CNOT gates <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 10,
         "elementary_gates" => ["RX_1", "RZ_1", "RZ_2", "CNot_1_2", "CNot_2_1", "Identity"],
-        "target_gate" => QCO.QFT2Gate(),
+        "target_gate" => QCOpt.QFT2Gate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "approximate",
         
@@ -398,21 +361,19 @@ function decompose_qft2_using_R()
         "RZ_discretization" => [-π/4, π/2, 3*π/4, 7*π/4],                      
         )
     
-        return params
 end
 
 function decompose_qft2_using_HT()
     println(">>>>> QFT2 Gate using H, T, CNOT gates <<<<<")
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
         "num_qubits" => 2, 
         "maximum_depth" => 8,
         "elementary_gates" => ["H_1", "H_2", "T_1", "T_2", "Tdagger_1", "Tdagger_2", "CNot_1_2", "CNot_2_1", "Identity"],
-        "target_gate" => QCO.QFT2Gate(),
+        "target_gate" => QCOpt.QFT2Gate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "exact", 
         )
     
-        return params
 end