+++ examples/5qubit_gates.jl
@@ -2,12 +2,12 @@ function decompose_RX_on_5qubits()
 
     println(">>>>> RX Gate on fourth qubit using U3Gate <<<<<")
  
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
     
     "num_qubits" => 5, 
     "maximum_depth" => 3,
     "elementary_gates" => ["H_1xCNot_2_3xI_4xI_5", "RX_2", "CU3_3_4", "Identity"], 
-    "target_gate" => QCO.kron_two_qubit_gate(5, QCO.CRXGate(π/4), "q3", "q4"),
+    "target_gate" => QCOpt.kron_two_qubit_gate(5, QCOpt.CRXGate(π/4), "q3", "q4"),
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
 
@@ -16,7 +16,5 @@ function decompose_RX_on_5qubits()
     "CU3_λ_discretization" => [0, π/2],
     "RX_discretization" => [π/2],
     )
-
-    return params
     
 end